requirements/base.txt
django == 5.1.1
python-environ == 0.4.54
psycopg2 == 2.9.9
django-ckeditor == 6.7.1
djangorestframework == 3.15.2
djangorestframework-simplejwt == 5.3.1
drf-spectacular == 0.27.2
celery == 5.4.0
redis == 5.1.0
django-filter == 24.3